requirements.txt
numpy>=1.23
pypdf>=4.0.0
sentence-transformers>=2.7.0
tqdm>=4.66
faiss-cpu>=1.7.4
requests>=2.31
matplotlib